Decouple pyramid and hfa.
